remove unnecessary store.
